fix for raspberry pi compiler using raspbian 6.3.0

+++ discrete_src/utils/simple_args.h
@@ -54,7 +54,7 @@ inline const char* SimpleArgs::get( int argn, char *argv[], int index, char* par
 		a = argn + a;
 	}
 
-	if ( a >= argn )
+	if ( (unsigned int)a >= (unsigned int)argn )
 	{
 		return 0;
 	}
